Added "update" process, many other adjustments

+++ ai_media_wizard/install_update/install.py
@@ -5,7 +5,6 @@ import stat
 from shutil import rmtree
 from subprocess import run
 
-from .. import options
 from .custom_nodes import install_base_custom_nodes
 
 LOGGER = logging.getLogger("ai_media_wizard")
@@ -36,25 +35,25 @@ def remove_readonly(func, path, _):
     func(path)
 
 
-def install(backend_dir="", flows_dir="", models_dir="") -> None:
-    """Performs clean installation."""
-    flows_dir = options.get_flows_dir(flows_dir)
-    if os.path.exists(flows_dir) is True:
-        LOGGER.info("Removing existing Flows directory: %s", flows_dir)
-        rmtree(flows_dir, onerror=remove_readonly)
-    os.makedirs(flows_dir)
-    models_dir = options.get_models_dir(models_dir)
-    if os.path.exists(models_dir) is True:
-        LOGGER.info("Removing existing Models directory: %s", models_dir)
-        rmtree(models_dir, onerror=remove_readonly)
-    os.makedirs(models_dir)
-    backend_dir = options.get_backend_dir(backend_dir)
-    if os.path.exists(backend_dir) is True:
-        LOGGER.info("Removing existing Backend directory: %s", backend_dir)
-        rmtree(backend_dir, onerror=remove_readonly)
-    os.makedirs(backend_dir)
-    run(f"git clone https://github.com/cloud-media-flows/ComfyUI.git {backend_dir}".split(), check=True)
-    run(f"python -m pip install -r {os.path.join(backend_dir, 'requirements.txt')}".split(), check=True)
-    with builtins.open(os.path.join(backend_dir, "extra_model_paths.yaml"), "w", encoding="utf-8") as fp:
-        fp.write(EXTRA_MODEL_PATHS.replace("amw_models_root", models_dir))
-    install_base_custom_nodes(os.path.join(backend_dir, "custom_nodes"))
+def install(backend_dir: str, flows_dir: str, models_dir: str) -> None:
+    """Performs a clean installation on the provided directories."""
+    if flows_dir:
+        if os.path.exists(flows_dir) is True:
+            LOGGER.info("Removing existing Flows directory: %s", flows_dir)
+            rmtree(flows_dir, onerror=remove_readonly)
+        os.makedirs(flows_dir)
+    if models_dir:
+        if os.path.exists(models_dir) is True:
+            LOGGER.info("Removing existing Models directory: %s", models_dir)
+            rmtree(models_dir, onerror=remove_readonly)
+        os.makedirs(models_dir)
+    if backend_dir:
+        if os.path.exists(backend_dir) is True:
+            LOGGER.info("Removing existing Backend directory: %s", backend_dir)
+            rmtree(backend_dir, onerror=remove_readonly)
+        os.makedirs(backend_dir)
+        run(f"git clone https://github.com/cloud-media-flows/ComfyUI.git {backend_dir}".split(), check=True)
+        run(f"python -m pip install -r {os.path.join(backend_dir, 'requirements.txt')}".split(), check=True)
+        with builtins.open(os.path.join(backend_dir, "extra_model_paths.yaml"), "w", encoding="utf-8") as fp:
+            fp.write(EXTRA_MODEL_PATHS.replace("amw_models_root", models_dir))
+        install_base_custom_nodes(os.path.join(backend_dir, "custom_nodes"))
